fix: 경매 및 낙찰 테이블 수정 및 등록시 상품명,결제기한 추가
summary:

footer:

+++ settlement/src/main/java/com/dev_high/settle/application/dto/SettlementRegisterRequest.java
@@ -2,12 +2,14 @@ package com.dev_high.settle.application.dto;
 
 import com.dev_high.order.domain.WinningOrder;
 
+import java.math.BigDecimal;
+
 public record SettlementRegisterRequest(
     String id,
     String sellerId,
     String buyerId,
     String auctionId,
-    Long winningAmount
+    BigDecimal winningAmount
 ) {
 
   public static SettlementRegisterRequest fromOrder(WinningOrder order) {
